requirements.txt
selenium
undetected-chromedriver
langchain-openai
langchain-core
python-dotenv
requests
beautifulsoup4